RakuAST: fully implement whatever currying of infixes
With this * * * + * * * results in a callable taking 4 args and
returning the sum of two products.

The difficulty here is that each tree node only knows its children but
not its parent. Thus a node cannot know whether there is a parent infix
that will take over the thunk. Therefore we need to call IMPL-CHECK on
the thunk right away to get names resolved and the stub for the code
object created. But the code object also references a signature. And
this signature can still change when the thunk gets expanded. So we need
to replace the signature in the code object whenever we add new
parameters to the thunk.