Fix for #5165
The problem was caused by not checking for object hashes when storing
a (object or normal) hash into another normal hash.

This takes another approach than Pull Request #5169, in that it does
*not* impose a penalty for *each* key that is being stored, (even in
normal hashes).  Instead, when a Mappy thing is encountered, a special
private method STORE_OBJECT_MAP is called that does the whole object
hash correctly.