RakuAST: improve canonicalization of ArrayComposer colonpairs
Custom operator lookup relies on the name to be canonicalized before
looking up. Thus infix:['<=»'] needs to become infix:<\<=»>
We already did this right for normal colonpairs, but array composers
still canonicalized to square brackets containing string literals.

I'm not sure if doing this for array composers containing a single
statement is enough, but it gets 03-cmp-ok.t passing.