Fix storing of typed Hashes from a list (#5172)
Prior to this commit, a Hash with typed keys ("hash object") couldn't be
properly stored in another Hash, typed or untyped, if it appeared in a
list. This is a regression from July 2016,
commit ecfb956bcf92488106a85f9462731095db11dfed .
In the affected timespan, the behavior was the following:

my %h{Any} = foo => 'bar', 12 => 'baz';
my %copy = (%h, );

The copy got recreated as the internal, `$!storage` representation.
Because of the current default implementation of `.hash`, the problem
appeared also when calling the `.hash` method on a list that contained a
"hash object", or enforcing hash context upon them with %().

STORE_AT_KEY is the method that takes care of creating an entry for an
implementation of a Hash, transforming the key and the value in a way the
underlying `$!storage` BOOTHash can hold them. When different
implementations of Hash are stored into another Hash (previously done with
the STORE_MAP method), this needs to be taken into account.

This commit fixes the problem by replacing non-dispatching STORE_MAP calls
with a PUSH_FROM_MAP method that needs to be provided for those
implementations of a Hash that have their own mapping to `$!storage`,
that is, typically the core implementations that have their own
STORE_AT_KEY. This PUSH_FROM_MAP knows how STORE_AT_KEY needs to be
called with regards to all of the elements retrieved from the Hash object
it's invoked on, meaning that all STORE_AT_KEY calls will happen with the
right arguments, at the cost of only one dispatch to the right PUSH_FROM_MAP
candidate.

Authored-by: Polgár Márton <ersterpleghthome@gmail.com>