RakuAST: un-break the REPL
The current REPL implementation loads the REPL class by running a comp
unit that's just "REPL", i.e. evaluates the class name. It does so by
stariting the compiler with :outer_ctx(nqp::null()). The old frontend
took this as a signal, that the comp unit's result is wanted. The
RakuAST frontend expected a concrete outer_ctx instead.

Fix by adopting the same criteria as the old frontend.