RakuAST: give slurpy placeholder params a fixed order in the signature
The old frontend added slurpy placeholder parameters to the signature in
the order of appearance in the block. At least one spectest relied on
this. In general we need compilation results to be reproducible.
However, we go through placeholders in hash order. That's not a problem
because positionals will be sorted lexicographically anyway. It was
a problem for slurpies, so just always add @_ before %_.