RakuAST: give named placeholder params a fixed order in the signature
The old frontend added named placeholder parameters to the signature in
the order of appearance in the block. In general we need compilation
results to be reproducible. However, we go through placeholders in hash
order. That's not a problem because positionals will be sorted
lexicographically anyway. It was a problem for nameds, so just give them
the same treatment as positionals.