RakuAST: fix smartmatch on $_ containing a regex boolifying result
Looking at the old compiler frontend, it seems like we don't need the
special smartmatch handling if the right hand side is a plain variable.
While it worked in most cases anyway, things got confusing when that
variable was $_. Then temporarization of that variable disturbed the
check whether the right hand side was a Regex (which doesn't want its
result to be boolified). Avoid the whole thing by just compiling to
a call to infix:sym<~~> for variables. Just like the old frontend.