Try to be more explicit when exception reporting dies
Currently when this happens at compile time we end up with useless 'No
exception handler found for catch' message providing us with no
information as to what the original cause was. With this commit we would
try to pull out as much info as possible at report it back with no bells
and whistles.