RakuAST: fix miscompilation of statements with both condition and loop modifiers
Postfix conditions are statement modifiers. They wraps the statement's
expression in the appropriate op (e.g. nqp::if) with appropriate return values.
A postfix for is also a statement modifier. It thunks the expression and wraps
that with appropriate looping code. The problem was when we had both of them on
a statement. Then the condition wrapper would have to go on the expression,
i.e. it must be part of the thunk. Instead we wrapped the calling of the thunk.
This was a problem because a for loop is just a call to .map on the list. This
map gets passed the thunk to run. Previously instead we evaluated the nqp::if
and passed the result to the map which got totally confused.

Fix by wrapping the expression in a condition thunk if we have both
condition and loops. However the condition thunk is only a pseudo thunk,
i.e. it does not actually wrap the expression in a code block. It's
merely used to wrap the expression in the condition before it gets added
to the loop modifier's thunk.

This solution is everything but pretty. It abuses the thunking mechanism
and may cause other problems in the future. However currently it's the
best I could come up with, without changing the whole architecture, so
let's give it a try.