Fix recent regression on storing Maps The commit that was meant to fix storage of "hash objects" into Hashes had a serious thinko: after checking for Maps, it invoked a method defined on Hashes only. This meant that after the changes
my %problem is Map = :foo;
(%problem,).hash

was failing, unable to find the PUSH_FROM_MAP method on Maps. (Ironically,
even the name suggested that it should be defined on Maps.)

This commit addresses this problem by indeed moving the PUSH_FROM_MAP method
to Map. For clarity, I added a type constraint to the parameter - the pushing
always happens *to* a Hash, that was the dedicated purpose of the method all
along.