Fix rare deadlocks during supply setup
In https://github.com/rakudo/rakudo/commit/26a9c313297a21c11ac30f02349497822686f507
a mechanism to solve deadlocks during supply / whenever setup was
introduced. In that implementation whenever a lock during setup
was hit a continuation would be taken and put into a list for later
processing. That list was then processed in order. If new locks were
encountered they were appened to the list. I was possible to end up
in a situation where the first continuation waited for a lock the
second continuation in the list held.

Later on in https://github.com/rakudo/rakudo/commit/547839200a772e26ea164e9d1fd8c9cd4a5c2d9f
a different mechanism to deal with deadlocks caused by recursion during
supply / whenever processing was introduced. That mechanism is far
simpler and neither involves continuations nor overwriting `$*AWAITER`.

It seems (this is not proven) that a precondition for deadlocks to
appear during supply / whenever setup is recursion locking. So the
second simpler mechanism is able to replace the first. This change does
just that. The first mechanism is removed leaving only the second.

A positive side-effect (appart from solving the rare deadlocks) is
that the whenever setup process is simpler and easier to reason about.
Especially locks unrelated to the supplies do not influence the setup
process anymore.