[JVM] Set Xmx4G for building Rakudo
As explained in https://github.com/rakudo/rakudo/issues/5186 the
compilation of blib/Perl6/BOOTSTRAP/v6c.jar failed with

  java.lang.OutOfMemoryError: Java heap space

otherwise. Hopefully we'll find a way to reduce the memory usage
for that step, because 4 GB is quite an amount of memory. (And
for instance, I could imagine, that we won't have that much memory
for the CI builds.)