RakuAST: fix resolver in dynamically compiled code missing symbols
The resolver used for resolving symbols in dynamically compiled code
needs to be the resolver as it was at the time when we created the stub
for that piece of code, because the resolver's state then reflects the
lexical environment that the code is defined in.

So use a deep clone of the resolver in the "lexical not found" callback.