RakuAST: Fix dynamically compiled blocks keeping compiler alive
When we compile a code object, we first "stub" it, i.e. create the meta object
(Sub, Method, Block, etc.) and assign a placeholder code to its $!do property.
This placeholder when executed will then "dynamically" compile the actual body
and replace itself on its code object.

This body could contain nested code objects like in the role example above
(the role body itself is a Block that's executed when the role is composed into
a class).  Since we're already compiling these nested code objects as part of
their parent body, we take the opportunity of "fixing up" their code objects
while we're at it, i.e. when Foo body is compiled, we also replace foo's $!do.

Normally when we load a comp unit, we fix up all routines, i.e. assign the
compiled blocks to their owning objects $!do, thereby replacing all stubs. But
if a stub was already replaced, that's not necessary, so we don't bother.
Except that there is this subtle difference between the dynamically compiled
bodies and the normal ones: the dynamically compiled code belongs to a
different comp unit (the one created for compiling that thing while the outer
comp unit is still unfinished). And this outer comp unit doesn't contain all
lexical symbols.

The same situation exists on the old frontend as well, except for that very
last bit: on the old frontend we create a wrapper frame containing all lexical
symbols visible at the point of the routine's definition. This includes all
symbols from the setting as well, so these frames are very large.  In the
RakuAST frontend we instead use a newly introduced mechanism of the VM:
fallback resolvers. I.e. when the VM can't find a lexical, it runs a callback
which will then access the RakuAST::Resolver. Since this fallback resolver is
a property of the comp unit, the same thing actually happens at runtime. I.e.
these dynamically compiled routines actually keep the resolver object alive
and that in turn references a lot of AST nodes. So definitely not memory
friendly. Also not necessary, since at runtime we'll have a complete lexical
environment there already.

Just do the normal fixup at load time, even for dynamically compiled blocks.