Switch Supply.zip to a watermark approach
In https://github.com/Raku/problem-solving/issues/364 a proposed
solution to a deadlock in supply setup is hindered by the failure of
a spectest for Supply.zip.

Thus far we defined Supply.zip as having its reuslt Supply be `done` as
soon as any input is `done`. This makes its behavior highly sensitive to
the timing of its input values; as a reuslt, it may not emit all of the
complete tuples of values that it feasibly could.

This change introduces a "watermark", which is the maximum number of
tuples we could possibly emit. When any input `supply` becomes done, we
take the number of messages that were delivered, and use that as the
maximum possible number of tuples that we can emit. Any further `done`s
might lower that watermark. The `supply` block as a whole is done when
all input supplies reach the watermark.

This passes all current spectests. Hopefully, it will also pass when
used in combination with the deadlock fix.