RakuAST: support non-constant initializers for attributes
This generalizes support for attribute initializers, so in addition to
has $!foo = 1; we can do has $!foo = rand;
However, it currently does so at the cost of generating runtime calls
even for compile time constants. To get the latter we'd need a way to
determine whether an expression's result is known at compile time.