Fix raciness in supply processing
Supply blocks are blocked by a lock from being entered in parallel.
During recursion, which can happen when supply blocks directly
contain `emit`s and with nested `whenever`s, the
`Lock::Async::protect-or-queue-on-recursion` method queues the code
to be run after the supply block is available again. That code is then
put on the thread pool. This introduces a race condition where the
code following the `emit` that triggered the recursion can be run
before the emit is processed entirely.

To solve this, the queuing logic is changed, to synchronously call the
queued code objects. This has several consequences:

- The queued code is guaranteed to be done once the line of code
  releasing the lock has finished. So there is no raciness anymore.
- The setup code is reproducible, so if it runs once, it will always run
  as the order of the queued code blocks does not change between runs.
- Dead locks caused by recursion during the setup logic are provably
  ruled out. Because the setup code has a single code path, locks can't
  be hidden in continuations or other threads. The
  `Lock::Async::protect-or-queue-on-recursion` guarantees that
  deadlocks by recursion don't happen. It previously couldn't do so
  reliably, because the recursion wasn't guaranteed to stay on the
  same call stack (continuations could move them out).
- In 547839200a772e26ea1 the recursion breaking logic and
  `Lock::Async::protect-or-queue-on-recursion` was originally
  implemented. The commit message states: "A holder's recursion
  competes fairly with outside messages, thanks to the queueing
  being through Lock::Async." This is not true anymore, as the queued
  code objects are now synchronously called and not put on the thread
  pool anymore.