Provide better error message for my %h = Callable
If it is the *only* value stored in a Map, and it is a Callable,
provide an error message that points to the use of ; vs , in the
contents of the Callable ({ :a; :b } versus { :a, :b }).

Addresses https://github.com/rakudo/rakudo/issues/5170

This change may actually be a performance improvement, as it
reduces the binary size of the STORE methods involved, making
it more likely they can be inlined.