RakuAST: make redeclarations a typed worry / panic
Until now, with RakUAST, lexical entries that were defined more than
once, would fail at the MAST generation stage.  Now, they will
properly worry / panic at the stage when being parsed.

This is achieved by having the .declare-lexical return a Bool that
indicates whether the name of the declaration already existed or not.
And then add the worry / throw the panic if it did exist already.

This does *not* however completely mimic the behaviour of the old
grammar, as in the case of a single worry, MAST generation still
complains of finding duplicate lexical symbols.  So I guess this
needs more investigation.