Allow EVAL of RakuAST::Nodes without MONKEY
Since there is no danger of injection with RakuAST::Node, there's
no need for "use MONKEY-SEE-NO-EVAL" if you're EVALling RakuAST nodes.

Originally, this was a compile time error.  Now it is a run-time error,
because we cannot see the difference between a string and a RakuAST
node at compile time.

Runtime check is achieved by setting an additional named argument
"no-monkey" in the call to EVAL that will be set to True if no
"use MONKEY-SEE-NO-EVAL" was seen at compile time.  The EVAL sub
then checks for that flag and throws if it is true and what is
being EVALled is not a RakyuAST node.