RakuAST: add deparsing/raku/test for RakuAST::Type::Setting
The tests are actually TODOd as the RakuAST::Type::Setting class
only does lexical lookups at the moment, rather than just looking
in the setting.

Also of note: there is currently **NO** syntax for lookups in the
setting that is also usable as type object when specifying constraints.

    my constant Int = Str;
    my Int $a = "foo";

does *not* create a type constraint error.  Maybe that's a good thing,
maybe not.  But what if we *would* like to specify the SETTING version
of Int?

    my SETTING::<Int> $a;
    ===SORRY!===
    Could not locate compile-time value for symbol SETTING

Perhaps we should consider the type slot as a BEGIN time thunk that
would know how to handle SETTING::<> PseudoStash lookup?