RakuAST: make blocks embedded in thunks nest properly
In "yes" andthen "{dd $_}" the "{dd $_}" gets compiled to a block and a call
to that block. The andthen thunks its RHS. For this to work, the block that's
contained in the string must appear inside the thunk's block. And that
although an expression is not a scope by itself.

Fix by making expressions ImmediateBlockUsers if they are thunked and
then adding block declarations to the thunk block.