RakuAST: correct .raku/deparse for roles
A Rakuast::Package that is a role actually converts its body into
a Sub that returns some pertinent info when run (which happens when
a role is consumed).  This commit hides that code when you do a .raku
or deparse on such a Package object.  The actual code in there can
still be seen if you .raku/deparse the body of the Package directly.

Unfortunately, this does not allow .raku roundtripping yet, as the
twiddling of the role body is currently done in the Actions, whereas
that should really be part of the RakuAST::Package object creation
if the package-declarator is "role".