RakuAST: use Signature.parameters-initialized as rendering check
And remove RakuAST::PlaceholderParameterOwner.code-has-placeholders
logic.  The latter suffered from the following situation:

    sub foo (@_) { @_ }

This would .code-has-placeholders return True, because the @_ inside the
scope, is codegenned as a RakuAST::VarDeclaration::Placeholder::SlurpyArray.
Yet, clearly this sub has a signature and @_ is not a slurpy at all.
Fortunately, that just QASTs correctly to a lexical lookup for '@_'.

Using .parameters-initialized as tests, has some effects on deparsing,
so some tests needed to be adapted.

This also made clear that there is no Raku syntax for a nameless sub
that has traits *and* placeholders.  Yet, one can create such a sub
easily using RakuAST calls.  This will then deparse as unparseable
code.  You can't have it all.