RakuAST: first stab at deparsing/.raku of regex/token/rule
The main problem was that "regex" / "token" / "rule" all codegenned
to RakuAST::RegexDeclaration, with (probably) tweaks applied to the
body.

This commit introduces two subclasses of RakuAST::RegexDeclaration:
TokenDeclaration and RuleDeclaration, and adds a "declarator" method
to all three.  By altering the $*SCOPE_KIND accordingly in the
grammar, the correct class is codegenned, without any other functional
changes.

This does *not* address the alterations made to the body yet: this
will require further research and most likely, modifications, as
these alterations would need to be done in the RakuAST classes, and
not in grammar or actions.