RakuAST: add .raku support for RakuAST::RuleDeclaration
This one is a little more troublesome, as it needs to remove the
RakuAST::Regex::WithSigspace entries from the tree before an actual
.raku representation can be made.

This also adds a RakuAST::Regex::QuantifiedAtom.replace-atom method
as the atom can also have a RakuAST::Regex::WithSigspace in it.

Also make sure RakuAST::Regex::WithSigspace correctly deparses with an
additional space behind it.