RakuAST: rework deparsing of regexes
Since a trailing space on an atom can have meaning when :sigspace
is active, we cannot deparse regexes always with spaces around them.
This is a conservative approach that may change in the future if
we've found another way to make sure that the semantics of a deparsed
regular expression does not differ from the AST.

This required a lot of changes with tests.