RakuAST: rework how regular expressions are codegenned
Before this commit, the behaviour of :s (aka sigspace) was codegenned
into the ASTs.  After this commit, the sigspace functionality is a
a parameter taken into account when creating the QAST for a regex.

This means that as a developer building custom regex ASTs, one does
not need to indicate sigspace functionality explicitely when building.

This is achieved by introducing a RakuAST::Regex::WithWhitespace class
that wraps around any regex term/atom if it is to be considered followed
by whitespace.  So / foo/ would codegen to a Regex::Literal, whereas
/ foo / would codegen to a Regex::WithWhitespace that wraps a Regex::Literal.
The wrapper will either produce QAST for the wrapped term/atom, or a
concat with <ws> if the :sigspace setting was active at moment of QAST
creation.

An overview of the changes:

- The RakuAST::Regex::WithSigspace class has effectively been renamed to
  RakuAST::Regex::WithWhitespace, and the QAST production has been adapted
  to honour the "s" flag.
- QAST production of the RakuAST::RegexDeclation base class now sets the
  "s" flag when producing for a rule (aka RakuAST::RuleDeclaration).
- A method "whitespace-wrappable" has been added to RakuAST::Regex::Term
  to indicate whether the term/atom can be wrapped with ::WithWhitespace.
- The specialized RakuAST::RuleDeclaration.raku method has been removed,
  as its special handling of the AST is no longer needed.
- Some RakuAST tests that checked sigspace handling, needed adaptation

This now allows regex ASTs to be reused without any need for adaptation
should the behaviour of sigspace need to be changed.