RakuAST: one step closer to supporting "no strict"
- create a subclass of Var::Lexical called Var::Lexical::Auto
- have the Action.compile_variable_access check for "no strict" and
  create a Var::Lexical::Auto instead of Var::Lexical
- have the Grammar.check_variable method only check exact Var::Lexical
  objects, and panic if not defined
- add Var::Lexical::Auto.resolve-with method that *if* it resolves
  just carries on.  If it does *not* resolve, then for now dies.

Currently lacking the RakuAST fu to do the right things to define
an "our" variable and mark it as resolved.  Hoping that nine or someone
else sees this as ultra LHF and can replace the nqp::die with the
correct code.

This approach should not produce any overhead for the normal path of
handling lexical variable references.