RakuAST: another step closer to making "no strict" work
This generates the correct QAST for auto-declared variables.  But to
prevent the QAST to generate this for every occurrence of an auto-
declared variable, it needs to mark it as declared.

The resolver has a "declare-lexical" method to do just that.  But only
if the resolver is not in "batch" mode.  And so far it always seems to
be.  Even if the declaration is happening on a CHECK-BEGIN method.  So
I'm unclear on how to proceed, and am hoping nine will be able to produce
some clarity.