Introduce Pod::Declarator class
This is basically the Pod::Block::Declarator class, but no longer a
subclass of Pod::Block, as Pod declarators are not Pod::Blocks as
per S26.  This loses two unused attributes per pod declarator.

Created a Pod::Block::Declarator class as a subclass of Pod::Declarator
to remain source compatible.