RakuAST: WIP on RakuST::Pod::Declarator class
Added basic class with functionality for creation/updating, along
with deparsing and .raku.  Ran into a but of a conceptual issue on
where to qast the declarator objects: in the legacy grammar, it appears
they're being qasted into $=pod, but with RakuAST classes one shouldn't
need to have that.

This will require some more thought.  Committing this work so far to
make sure it's visible and can be commented upon.