RakuAST: introducing RakuAST::Doc modules
These provide the building blocks for creating a simplified
representation of rakudoc (aka pod).  For now, this only includes
creation, deparsing and .raku + tests.  The following classes are
provided: ::Block, ::Paragraph, ::Verbatim, ::Table, ::Markup

- RakuAST::Doc::Block
Represents a pod block, with type, level, config and ::Doc::Paragraph
objects

- RakuAST::Doc::Paragraph
Represents an actual paragraph with text atoms (usually a combination
of ::StrLiterals and ::Doc::Markup objects)

- RakuAST::Doc::Verbatim
Represents a verbatim pod block, such as with =comment and =code
and the speculated "ambient" block.  With type, level, config and
::StrLiteral text

- RakuAST::Doc::Table
Represents a doc table, with type (by default "table"), level,
config, headers and rows.  Exact semantics of headers and rows still
to be decided.

- RakuAST::Doc::Markup
Represents a doc markup sequence, consisting of a letter (e.g. "B"
for bold), one or more atoms (usually a combination of ::StrLiterals
and ::Doc::Markup objects), optional meta information (e.g. the URL
in case of "L", and one or more lemmas with "X") and a separator
to be used to deparse multiple lemmas (default: empty string)

The next step will be adapting the Raku grammar to be producing these
RakuAST objects for rakudoc (pod).  The step after that will be to
handle the QAST production of these objects.