RakuAST: allow doc blocks in statement lists
It is now possible to put RakuAST::Doc::Blocks in statement lists,
and still have them EVAL correctly, as if they're not there.  This
is achieved by checking all of the statement list logic for doc blocks
and omitting them for further handling.

This approach appeared superior to an approach with a proper
RakuAST::Statement::Doc object, because that would make it much
harder to make a doc block invisible, especially at the end of a
scope, without altering the return value.  It would also make it
a proper statement, which it isn't  :-)

Also make "abbreviated" a proper RakuAST::Doc::Basic flag, without
needing to interfere with the other config.

Add / adapt tests RakuAST tests accordingly.