RakuAST: Add RakuAST::Doc::DeclaratorTarget.add-leading/trailing
To facilitate adding leading / trailing declarator comments from
the grammar.  Calling add-leading/trailing on a DeclaratorTarget
such as a RakuAST::Sub, will automatically create a Declarator
object and linke that to the target, and vice-versa, and add the
given documentation.  If there already was a Declarator object,
then it will just simply add the given documentation to that.

Also add test.