RakuAST: fix .raku of Rakuast::Doc::DeclaratorTarget objects
Creating a .raku representation of object 1 that has an object 2 as
an attribute, that has object 1 as an attribute, it is a bit of a
challenge to prevent infinilooping.

This adds a "declarator-docs" method to DeclaratorTargets, which
takes :leading and :trailing, creates a RakuAST::Doc::Declarator
target with that and for the WHEREFORE to the invocant.  It then
sets that in the WHY of the invocant, and returns the invocant.

Consequently, the .raku methods of DeclaratorTargets are adapted
such that they will create a call to the .declarator-docs method
at the end of the rakufication of the rest of the object.