RakuAST: make ENTER phaser results thread safe
We must not share the Scalar the result of an ENTER phaser gets stored
in between callers. Creating a Scalar at compile time works for INIT
phasers because they are guaranteed to only run once in the life time of
a program. ENTER can be called multiple times and in parallel, so we
need to store the result in a local variable instead.