RakuAST: complete grammar support for declarator docs
Trailing declarator docs can now only be specified if they start on
the same line as the declarand they refer to.  This is a much saner
approach than in the legacy grammar where it was quite uncertain
where (if ever) trailing declarator docs would get attached to.

    #| leading
    sub a { #= single line trailing
    }

    #| leading
    sub a { #=(multi
    line
    trailing)
    }

Any trailing declarator doc that can not be attached to a declarand
will now add a typed worry.

Plus some extensive rework on deparsing of anything that can have
declarator docs.

Currently the grammar entries for bare { } and -> { } do not seem
to trigger yet.  Those tests are marked as todo for now.  It's not
like I've seen declarator docs on bare / pointy blocks in the wild.