RakuAST: QAST $=finish as an SVal instead of a WVal
This makes $=finish to be properly available after pre-compilation.

Also note that in the legacy grammar, $=finish was actually codegenned
as a lexical variable in the compunit.  In the Raku grammar, it is
codegenned as an SVal *without* the $=finish lexical variable actually
existing (a behaviour similar to e.g. $?FILE and $?LINE).  Since
$=finish would generally only be called once in a program anyway,
I don't think there'll be a lot of doubling.  On the plus side: if
some code has a large =finish block, that will now *not* actually
be codegenned if there is no reference to $=finish.