Introduce Str.htmlparse
This will try to interprete the string as an HTML entity, and return
the string associated with it, for example: &aacute;".htmlparse -> á

This introduces a rather large lookup table, which in fact already
lives in src/Perl6/Pod.nqp.  But since all src/Perl6 files are on
the way out in the RakuAST world, it felt that it might as well be
updated and made generally accessible, just as Str.uniparse.

The & prefix and ; postfix are optional in the lookup.

This adds two source files: one generator file which creates the
source code for the lookup table, and a source file to be included
in the setting (augmenting the Str class with the "htmlparse" method).

Running the generator script will fetch the most up-to-date entity data
from the Net (currently: https://html.spec.whatwg.org/entities.json)