Introduce Int.html-entity
Attempts to return the HTML entity associated with the integer value
(assumed to be a codepoint).  Returns Nil if there is no such entity.

Lazily builds the lookup hash from the Str.htmlparse lookup table
the first time the Int.html-entity method is invoked, so this won't
increase build/startup time significantly, while being able to
provide a very nice feature when it is needed.