Add Str.leading/trailing-whitespace implemementation-detail
Returns any leading / trailing whitespace of the string, which would
be the empty string if there is no leading/trailing whitespace, or
the string itself if it only consists of whitespace or is empty.

The Str.leading-whitespace will be used in RakuAST rakudoc block
parsing and codegenning.  Str.trailing-whitespace added for consistency.

Both are marked implementation-detail for now, unless we agree on
making it a public API.