RakuAST: simplify Doc::Block hierarchy
While prototyping the Rakudoc slang, it occurred to me that the
RakuAST::Doc::Block hierarchy was already too refined, and that the
eventual creation of Pod::xxx objects could be done from one type
of RakuAST::Doc::Block class.  This effectively renames the
RakuAST::Doc::Formatted class to RakuAST::Doc::Block, and removes
the RakuAST::Doc::Verbatim and RakuAST::Doc::Table classes.

This will allow more of the pod parsing and processing to be done
in Raku, rather than in NQP.