RakuAST: add RakuAST::Doc::Block.from-paragraphs method
This method will inspect the paragraphs as collected by the Doc
grammar an produce the correct contents depending on the type of
block.  This includes automatic code block detection, as well as
any markup creation.

This effectively replaces a large part of the old Pod grammar.

Also tweak some of the legacy pod creation logic.