RakuAST: make-legacy-pod for =table NYI for now
The legacy grammar would do its best to interpret =table specifications
as best as possible in NQP.  The new grammar just collects the
paragraph(s) of the =table specification, leaving it to the legacy
pod generation to make sense of.  But that code hasn't been written
yet, so make it a NYI for now.