RakuAST: Report illegal postdeclaration of normal variables
Makes my $i; { $i++; my $i; } fail properly

This replaces the mechanism previously used for finding illegal post
declarations of dymanic variables. Instead of relying on the order that
nodes are attached to the tree, we instead piggy back on the traversal
that already happens for finding lexical declarations. We simply record
the order of declarations and accesses and then try to find any
declaration that happens after the access.

This however requires that the traversal of the tree happens in depth
first order and that nodes visit their children in lexical order, too.